+++ lib/detailed_page.dart
